requirements.txt
peewee==3.16.0
Flask==2.1.4
Flask-Script==2.0.9
Flask-WTF==1.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==3.0.0
APScheduler==3.8.0
PyJWT==2.1.0
json5==0.9.6
html5lib==1.1
requests==2.26.0
PyMySQL==1.0.2
SQLAlchemy==1.4.22
